requirements.txt
WMI
psutil
pystray
requests
pywin32
PyAutoGUI
keyboard
PyGetWindow
elevate
plyer
comtypes
tkcalendar
Pillow
pynput
pycaw
openai
google-generativeai
